+++ docs/CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - Dropped support for python 3.9
 
+### Fixed
+
+- Made Singleton metaclass thread-safe
+
 ## [0.10.0] - 2024-12-09
 
 ### Added